inline single-use get-workspace-nodes function
